requirements.txt
pandas==2.3.2
requests==2.32.5
beautifulsoup4==4.13.5
python-dotenv==1.1.1
selenium==4.35.0
